+++ CHANGELOG.md
@@ -14,9 +14,15 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- richat: impl subscribe_jup ([#140](https://github.com/lamports-dev/richat/pull/140))
-
 ### Breaking
+
+## 2025-10-06
+
+- richat-v5.1.0
+
+### Features
+
+- richat: impl subscribe_accounts ([#140](https://github.com/lamports-dev/richat/pull/140))
 
 ## 2025-09-17